requirements.txt
fastapi
uvicorn
torch
torchaudio
chatterbox-tts
websockets
python-multipart
requests